requirements.txt
certifi==2019.3.9
chardet==3.0.4
click==7.1.2
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
idna==2.8
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.1.1
requests==2.28.1
SQLAlchemy==1.3.1
Werkzeug==0.14.1
bcrypt==4.0.1
boto3==1.16.25
botocore==1.19.25
jmespath==0.10.0
Pillow==8.0.1
python-dateutil==2.8.1
s3transfer==0.3.3
six==1.15.0
urllib3==1.25.4
python-dotenv==0.21.0
